細かい説明の追加や必要のないコメントの削除
